requirements.txt
playwright
torch==2.1.2
xformers==0.0.23.post1
crawl4ai[all] @ git+https://github.com/unclecode/crawl4ai.git
streamlit_chat 
numpy>=1.26.0,<3
typing_extensions
langserve
fastapi
httpx
langgraph
setuptools==75.6.0
absl-py==2.1.0
accelerate==0.25.0
aiofiles>=24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.9.5
aioprometheus==23.3.0
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anthropic==0.30.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attributedict==0.3.0
attrs==24.2.0
Authlib==1.3.0
Babel==2.14.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blessings==1.7
blinker==1.7.0
boto3==1.34.142
botocore==1.34.142
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
chevron==0.14.0
click==8.1.7
cmake==3.27.1
codecov==2.1.13
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.2.1
contourpy==1.2.0
coverage==7.4.1
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.5.14
DataProperty==1.0.1
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.30.1
dill==0.3.7
distlib==0.3.8
distro==1.7.0
docstring_parser==0.16
duckduckgo_search==6.1.9
entrypoints==0.4
eval_type_backport==0.2.0
evaluate==0.4.1
exceptiongroup==1.1.2
executing==2.0.1
Faker==26.0.0
fastapi==0.111.0
fastapi-cli==0.0.5
fastavro==1.9.5
favicon==0.7.0
feedparser==6.0.11
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
FlagEmbedding==1.2.10
torch
Flask
Flask-Cors
Flask-RESTful
flask-restful-swagger-2
flatbuffers==24.3.25
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
htbuilder==0.6.2
html2text==2024.2.26
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.4
ijson==3.3.0
immutables==0.20
importlib-resources==6.1.1
importlib_metadata==7.1.0
inspecta==0.1.3
instructor==1.4.0
ipykernel==6.28.0
ipython==8.19.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jiter==0.4.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
json_repair==0.8.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema>=4.22.0,<5.0.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.2.7
langchain-anthropic==0.1.19
langchain-community==0.2.7
langchain-core==0.2.12
langchain-experimental==0.0.62
langchain-google-vertexai==1.0.6
langchain-groq==0.1.6
langchain-text-splitters==0.2.2
langgraph==0.1.5
langserve==0.2.2
langsmith==0.1.83
launchpadlib==1.10.16
Lazify==0.4.0
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
limits==3.13.0
lit==16.0.6
litellm>=1.53.1
literalai==0.0.607
lm_eval==0.4.1
lxml==5.3
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.12
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nh3==0.2.15
ninja==1.11.1
nltk==3.8.1
notebook_shim==0.2.3
numexpr==2.9.0
oauthlib==3.2.0
onnxruntime==1.18.0
openai>=1.45.0,<2.0.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
orjson==3.10.5
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
pathvalidate==3.2.0
pdf2image==1.17.0
peft==0.7.1
pexpect==4.9.0
pillow==10.4
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.4.0
portalocker==2.8.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.11.1
pycparser==2.21
pydeck==0.9.1
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.8.1
pyngrok==7.2.0
pynvml==11.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyproject-api==1.6.1
pyproject-toml==0.0.10
pyre-extensions==0.0.29
pyreqwest_impersonate==0.4.9
referencing==0.30.2
regex==2023.8.8
requests==2.32.3
responses==0.18.0
rootpath==0.1.1
rouge-score==0.1.2
rpds-py==0.9.2
rsa==4.9
s3transfer==0.10.2
sacrebleu==2.4.0
safetensors==0.4.1
scikit-learn==1.5.0
scipy==1.12.0
SecretStorage==3.3.1
selenium==4.23.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
sgmllib3k==1.0.0
shapely==2.0.4
shellingham==1.5.4
shortuuid==1.0.11
simple-websocket==1.0.0
six==1.16.0
slowapi==0.1.9
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlitedict==2.1.0
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
svgwrite==1.4.3
sympy==1.12
syncer==2.0.3
tabledata==1.3.3
tabulate==0.9.0
tavily-python==0.3.3
tcolorpy==0.1.4
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.0
texttable==1.7.0
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.2.1
together==1.2.12
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tox==4.12.1
tqdm==4.66.5
tqdm-multiprocess==0.0.11
traitlets==5.14.1
transformers==4.41.2
Tree==0.2.4
types-python-dateutil==2.8.19.14
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.30.1
uvloop==0.19.0
validators==0.34.0
virtualenv==20.25.0
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.0.1
weaviate-client==4.9.6
wikipedia==1.4.0
Wikipedia-API==0.6.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==1.0.0
pydantic>=2.8.0,<3.0.0
